core: envelopes: fix bug with capacitySpeedLimit not stopping the train
